+++ include/ensmallen_bits/cmaes/cmaes_impl.hpp
@@ -136,6 +136,9 @@ typename MatType::elem_type CMAES<SelectionPolicyType,
 
   BaseMatType transformedIterate = transformationPolicy.Transform(iterate);
 
+  // Controls early termination of the optimization process.
+  bool terminate = false;
+
   // Calculate the first objective function.
   ElemType currentObjective = 0;
   for (size_t f = 0; f < numFunctions; f += batchSize)
@@ -145,8 +148,8 @@ typename MatType::elem_type CMAES<SelectionPolicyType,
         effectiveBatchSize);
     currentObjective += objective;
 
-    Callback::Evaluate(*this, function, transformedIterate, objective,
-        callbacks...);
+    terminate |= Callback::Evaluate(*this, function, transformedIterate,
+        objective, callbacks...);
   }
 
   ElemType overallObjective = currentObjective;
@@ -174,12 +177,9 @@ typename MatType::elem_type CMAES<SelectionPolicyType,
   // The current visitation order (sorted by population objectives).
   arma::uvec idx = arma::linspace<arma::uvec>(0, lambda - 1, lambda);
 
-  // Controls early termination of the optimization process.
-  bool terminate = false;
-
   // Now iterate!
-  terminate |= Callback::BeginOptimization(*this, function, 
-      transformedIterate, callbacks...);
+  Callback::BeginOptimization(*this, function, transformedIterate,
+      callbacks...);
 
   // The number of generations to wait after the minimum loss has
   // been reached or no improvement has been made before terminating.
@@ -217,7 +217,8 @@ typename MatType::elem_type CMAES<SelectionPolicyType,
 
       // Calculate the objective function.
       pObjective(idx(j)) = selectionPolicy.Select(function, batchSize,
-        transformationPolicy.Transform(pPosition[idx(j)]), callbacks...);
+          transformationPolicy.Transform(pPosition[idx(j)]), terminate,
+          callbacks...);
     }
 
     // Sort population.
@@ -231,7 +232,8 @@ typename MatType::elem_type CMAES<SelectionPolicyType,
 
     // Calculate the objective function.
     currentObjective = selectionPolicy.Select(function, batchSize,
-      transformationPolicy.Transform(mPosition[idx1]), callbacks...);
+        transformationPolicy.Transform(mPosition[idx1]), terminate,
+        callbacks...);
 
     // Update best parameters.
     if (currentObjective < overallObjective)