+++ include/ensmallen_bits/katyusha/katyusha_impl.hpp
@@ -112,8 +112,7 @@ KatyushaType<Proximal>::Optimize(
 
   const size_t actualMaxIterations = (maxIterations == 0) ?
       std::numeric_limits<size_t>::max() : maxIterations;
-  terminate |= Callback::BeginOptimization(*this, function, iterate,
-      callbacks...);
+  Callback::BeginOptimization(*this, function, iterate, callbacks...);
   for (size_t i = 0; i < actualMaxIterations && !terminate; ++i)
   {
     // Calculate the objective function.
@@ -125,7 +124,8 @@ KatyushaType<Proximal>::Optimize(
           effectiveBatchSize);
       overallObjective += objective;
 
-      Callback::Evaluate(*this, function, iterate0, objective, callbacks...);
+      terminate |= Callback::Evaluate(*this, function, iterate0, objective,
+          callbacks...);
     }
 
     if (std::isnan(overallObjective) || std::isinf(overallObjective))
@@ -174,7 +174,7 @@ KatyushaType<Proximal>::Optimize(
     double cw = 1;
     w.zeros();
 
-    for (size_t f = 0, currentFunction = 0; f < innerIterations;
+    for (size_t f = 0, currentFunction = 0; (f < innerIterations) && !terminate;
         /* incrementing done manually */)
     {
       // Is this iteration the start of a sequence?
@@ -254,7 +254,10 @@ KatyushaType<Proximal>::Optimize(
           effectiveBatchSize);
       overallObjective += objective;
 
-      Callback::Evaluate(*this, function, iterate, objective, callbacks...);
+      // The optimization is finished, so we don't need to care about the
+      // callback result.
+      (void) Callback::Evaluate(*this, function, iterate, objective,
+          callbacks...);
     }
   }