+++ include/ensmallen_bits/sgd/sgd_impl.hpp
@@ -124,7 +124,7 @@ SGD<UpdatePolicyType, DecayPolicyType>::Optimize(
   BaseGradType gradient(iterate.n_rows, iterate.n_cols);
   const size_t actualMaxIterations = (maxIterations == 0) ?
       std::numeric_limits<size_t>::max() : maxIterations;
-  terminate |= Callback::BeginOptimization(*this, f, iterate, callbacks...);
+  Callback::BeginOptimization(*this, f, iterate, callbacks...);
   terminate |= Callback::BeginEpoch(*this, f, iterate, epoch,
       overallObjective, callbacks...);
   for (size_t i = 0; i < actualMaxIterations && !terminate;
@@ -181,9 +181,7 @@ SGD<UpdatePolicyType, DecayPolicyType>::Optimize(
         return overallObjective;
       }
 
-      if (std::abs(lastObjective - overallObjective) < tolerance ||
-          Callback::BeginEpoch(*this, f, iterate, epoch, overallObjective,
-              callbacks...))
+      if (std::abs(lastObjective - overallObjective) < tolerance)
       {
         Info << "SGD: minimized within tolerance " << tolerance << "; "
             << "terminating optimization." << std::endl;
@@ -191,6 +189,9 @@ SGD<UpdatePolicyType, DecayPolicyType>::Optimize(
         Callback::EndOptimization(*this, f, iterate, callbacks...);
         return overallObjective;
       }
+
+      terminate |= Callback::BeginEpoch(*this, f, iterate, epoch,
+          overallObjective, callbacks...);
 
       // Reset the counter variables.
       lastObjective = overallObjective;
@@ -202,8 +203,11 @@ SGD<UpdatePolicyType, DecayPolicyType>::Optimize(
     }
   }
 
-  Info << "SGD: maximum iterations (" << maxIterations << ") reached; "
-      << "terminating optimization." << std::endl;
+  if (!terminate)
+  {
+    Info << "SGD: maximum iterations (" << maxIterations << ") reached; "
+        << "terminating optimization." << std::endl;
+  }
 
   // Calculate final objective if exactObjective is set to true.
   if (exactObjective)
@@ -215,7 +219,9 @@ SGD<UpdatePolicyType, DecayPolicyType>::Optimize(
       const ElemType objective = f.Evaluate(iterate, i, effectiveBatchSize);
       overallObjective += objective;
 
-      Callback::Evaluate(*this, f, iterate, objective, callbacks...);
+      // The optimization is over, so it doesn't matter what the callback
+      // returns.
+      (void) Callback::Evaluate(*this, f, iterate, objective, callbacks...);
     }
   }